# docs/openapi2-oer-server-web.json
{
  "swagger": "2.0",
  "info": {
    "description": "oer-collector - JSON REST server backend for tv program data of public-law channels in Germany, Austria and Switzerland",
    "version": "v1",
    "title": "OER-Collector-Server - German public-law TV program data REST API",
    "contact": {
      "name": "emschu",
      "url": "https://github.com/emschu",
      "email": "look@github.com"
    },
    "license": {
      "name": "AGPLv3",
      "url": "http://www.gnu.org/licenses/agpl-3.0.de.html"
    }
  },
  "host": "localhost:8081",
  "basePath": "/",
  "tags": [
    {
      "name": "status-controller",
      "description": "Status Controller"
    }
  ],
  "paths": {
    "/v1/channel/{id}": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "channelSingleAction",
        "operationId": "channelSingleActionUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Channel"
            }
          },
          "500": {
            "description": "500 message",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/channels": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "channelListAction",
        "operationId": "channelListActionUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Channel"
              }
            }
          },
          "500": {
            "description": "500 message",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/program/daily": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "allDailyProgramEntries",
        "operationId": "allDailyProgramEntriesUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProgramEntry"
              }
            }
          },
          "500": {
            "description": "500 message",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/program/daily/{channelId}": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "dailyProgramEntriesByChannel",
        "operationId": "dailyProgramEntriesByChannelUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "description": "channelId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProgramEntry"
              }
            }
          },
          "500": {
            "description": "500 message",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/program/entry/{entryId}": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "programEntry",
        "operationId": "programEntryUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "description": "entryId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProgramEntry"
            }
          },
          "500": {
            "description": "500 message",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/program": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "Get program data",
        "operationId": "programEntriesUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "channelId",
            "in": "query",
            "description": "channelId",
            "required": false,
            "type": "integer",
            "format": "int64",
            "allowEmptyValue": false
          },
          {
            "name": "from",
            "in": "query",
            "description": "from",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "to",
            "in": "query",
            "description": "to",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProgramEntry"
              }
            }
          },
          "500": {
            "description": "500 message",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/status": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "statusAction",
        "operationId": "statusActionUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/StatusResponse"
            }
          },
          "500": {
            "description": "500 message",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "Channel": {
      "type": "object",
      "properties": {
        "adapterFamily": {
          "type": "string",
          "enum": [
            "ARD",
            "ZDF",
            "ORF"
          ]
        },
        "channelKey": {
          "type": "string",
          "enum": [
            "ARD",
            "ZDF",
            "ZDF_INFO",
            "ZDF_NEO",
            "DREISAT",
            "ARTE",
            "BR",
            "HR",
            "MDR",
            "NDR",
            "RBB",
            "RADIO_BREMEN_TV",
            "SR",
            "SWR_BW",
            "SWR_RP",
            "WDR",
            "ALPHA",
            "TAGESSCHAU_24",
            "ONE",
            "KIKA",
            "PHOENIX",
            "ORF1",
            "ORF2",
            "ORF2_EUROPE",
            "ORF_SPORT",
            "ORF3"
          ]
        },
        "homePage": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "technicalId": {
          "type": "string"
        }
      },
      "title": "Channel"
    },
    "Error": {
      "type": "object",
      "title": "Error"
    },
    "ImageLink": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "url": {
          "type": "string"
        }
      },
      "title": "ImageLink"
    },
    "ProgramEntry": {
      "type": "object",
      "properties": {
        "adapterFamily": {
          "type": "string",
          "enum": [
            "ARD",
            "ZDF",
            "ORF"
          ]
        },
        "channel": {
          "$ref": "#/definitions/Channel"
        },
        "createdAt": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "durationInMinutes": {
          "type": "integer",
          "format": "int32"
        },
        "endDateTime": {
          "type": "string"
        },
        "homePage": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageLinks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImageLink"
          }
        },
        "startDateTime": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tag"
          }
        },
        "technicalId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      },
      "title": "ProgramEntry"
    },
    "StatusResponse": {
      "type": "object",
      "properties": {
        "artistCount": {
          "type": "integer",
          "format": "int64"
        },
        "channelCount": {
          "type": "integer",
          "format": "int64"
        },
        "currentlyUpdating": {
          "type": "boolean"
        },
        "imageLinksCount": {
          "type": "integer",
          "format": "int64"
        },
        "programEntryCount": {
          "type": "integer",
          "format": "int64"
        },
        "serverDateTime": {
          "type": "string"
        },
        "tagCount": {
          "type": "integer",
          "format": "int64"
        },
        "tvChannels": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Channel"
          }
        },
        "tvShowCount": {
          "type": "integer",
          "format": "int64"
        },
        "version": {
          "type": "string"
        }
      },
      "title": "StatusResponse"
    },
    "Tag": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "tagName": {
          "type": "string"
        }
      },
      "title": "Tag"
    }
  }
}